Move map matching strategies into their own module
